feat(directoryservice): add new method to get resourceroom
