--- src/services/DirectoryService.ts
@@ -1,15 +1,0 @@
-import { DirectoryInfoReturn } from "types/directory"
-
-import { apiService } from "./ApiService"
-
-export const createResourceRoom = (
-  siteName: string,
-  { newDirectoryName, items }: DirectoryInfoReturn
-): Promise<void> => {
-  const endpoint = `/sites/${siteName}/resourceRoom/`
-  const body = {
-    items,
-    newDirectoryName,
-  }
-  return apiService.post(endpoint, body)
-}
